Renamed envs to env_atari.py and took out vizdoom part into env_vizdoom.py. Movement reward set to -0.001 and pickup euclidean distance to 0.3
