+++ algorithms/a3c/env_vizdoom.py
@@ -1,5 +1,5 @@
 """
-Adapted from: https://github.com/ikostrikov/pytorch-a3c/blob/master/envs.py
+Adapted from: https://github.com/devendrachaplot/DeepRL-Grounding/blob/master/env.py
 
 This contains auxiliary wrappers for the atari openAI gym environment e.g. proper resizing of the
 input frame and a running average normalisation of said frame after resizing
@@ -11,11 +11,10 @@ import collections
 import codecs
 import json
 
-import gym
 import numpy as np
-from gym.spaces.box import Box
+from gym import spaces
 
-from algorithms.a3c.vizdoom_utils.doom import *
+from algorithms.a3c.vizdoom_utils.doom import *   # todo probably avoid this but at least it is self-contained within this file
 from algorithms.a3c.vizdoom_utils.points import *
 from algorithms.a3c.vizdoom_utils.constants import *
 
@@ -24,6 +23,11 @@ actions = [[True, False, False], [False, True, False], [False, False, True]]
 ObjectLocation = collections.namedtuple("ObjectLocation", ["x", "y"])
 AgentLocation = collections.namedtuple("AgentLocation", ["x", "y", "theta"])
 
+
+class Task:
+    # extra addition to make it work with ai2thor code
+    def __init__(self):
+        self.task_has_language_instructions = True
 
 
 class GroundingEnv:
@@ -48,6 +52,15 @@ class GroundingEnv:
             self.get_all_objects(self.params.all_instr_file)
         self.object_sizes = self.read_size_file(self.params.object_size_file)
         self.objects_info = self.get_objects_info()
+
+        # extra stuff to make it work with ai2thor A3C
+        self.task = Task()
+        self.observation_space = spaces.Box(low=0, high=255,
+                                            shape=(3, self.params.frame_height,
+                                                   self.params.frame_width),
+                                            dtype=np.uint8)
+        self.action_space = spaces.Discrete(3)
+        self.task.word_to_idx = self.word_to_idx
 
     def game_init(self):
         """Starts the doom game engine."""
@@ -123,6 +136,8 @@ class GroundingEnv:
         reward = self.get_reward()
         is_final = False
         extra_args = None
+
+        print('New instruction: {}'.format(self.instruction))
 
         return state, reward, is_final, extra_args
 
@@ -354,59 +369,3 @@ class GroundingEnv:
                 if word not in word_to_idx:
                     word_to_idx[word] = len(word_to_idx)
         return word_to_idx
-
-
-# -----------------
-# Atari preprocessing and wrappers below
-# -----------------
-
-# Taken from https://github.com/openai/universe-starter-agent
-def create_atari_env(env_id):
-    env = gym.make(env_id)
-    env = AtariRescale42x42(env)
-    env = NormalizedEnv(env)
-    return env
-
-
-def _process_frame42(frame):
-    frame = frame[34:34 + 160, :160]
-    # Resize by half, then down to 42x42 (essentially mipmapping). If
-    # we resize directly we lose pixels that, when mapped to 42x42,
-    # aren't close enough to the pixel boundary.
-    frame = cv2.resize(frame, (80, 80))
-    frame = cv2.resize(frame, (42, 42))
-    frame = frame.mean(2, keepdims=True)
-    frame = frame.astype(np.float32)
-    frame *= (1.0 / 255.0)
-    frame = np.moveaxis(frame, -1, 0)
-    return frame
-
-
-class AtariRescale42x42(gym.ObservationWrapper):
-    def __init__(self, env=None):
-        super(AtariRescale42x42, self).__init__(env)
-        self.observation_space = Box(0.0, 1.0, [1, 42, 42])
-
-    def _observation(self, observation):
-        return _process_frame42(observation)
-
-
-class NormalizedEnv(gym.ObservationWrapper):
-    def __init__(self, env=None):
-        super(NormalizedEnv, self).__init__(env)
-        self.state_mean = 0
-        self.state_std = 0
-        self.alpha = 0.9999
-        self.num_steps = 0
-
-    def _observation(self, observation):
-        self.num_steps += 1
-        self.state_mean = self.state_mean * self.alpha + \
-            observation.mean() * (1 - self.alpha)
-        self.state_std = self.state_std * self.alpha + \
-            observation.std() * (1 - self.alpha)
-
-        unbiased_mean = self.state_mean / (1 - pow(self.alpha, self.num_steps))
-        unbiased_std = self.state_std / (1 - pow(self.alpha, self.num_steps))
-
-        return (observation - unbiased_mean) / (unbiased_std + 1e-8)
