+++ gym_ai2thor/config_files/NL_lookat_bowls_vs_cups_fp1_config.json
@@ -3,7 +3,7 @@
     "pickup_put_interaction": false,
     "lookupdown_actions": true,
     "max_episode_length": 1000,
-    "movement_reward": -0.01,
+    "movement_reward": -0.001,
     "scene_id": "FloorPlan1",
     "grayscale": true,
     "resolution": [128, 128],
